requirements.txt
# Core dependencies
streamlit>=1.28.0
streamlit-autorefresh>=1.0.1
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Interactive Brokers
ib-insync>=0.9.86

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
plotly>=5.17.0

# Utils
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0

# Data Sources
yfinance>=0.2.0

# Task Queue
celery>=5.5.0
redis>=7.0.0
